+++ nlgeometry/Writer/OffWriter.cpp
@@ -24,11 +24,13 @@
 #include <iostream>
 #include <fstream>
 #include <sstream>
+#include <unordered_map>
+
 
 namespace nlgeometry
 {
 
-  void OffWriter::writeMesh( MeshPtr mesh, const std::string& fileName_ )
+  void OffWriter::writeMesh( const MeshPtr mesh, const std::string& fileName_ )
   {
     Facets facets;
     facets.insert( facets.end( ), mesh->triangles( ).begin(),
@@ -38,7 +40,7 @@ namespace nlgeometry
     writeMesh( facets, mesh->vertices( ), fileName_ );
   }
 
-  void OffWriter::writeMesh( Facets& facets_, Vertices& vertices_,
+  void OffWriter::writeMesh( const Facets& facets_, const Vertices& vertices_,
                              const std::string& fileName_ )
   {
     std::ofstream outStream(fileName_.c_str());
@@ -50,20 +52,29 @@ namespace nlgeometry
 
     outStream << "OFF\n\n" << vertices_.size( ) << " " << facets_.size( )
               << " 0\n";
-    for ( unsigned int i = 0; i < ( unsigned int ) vertices_.size( ); i++ )
+    std::unordered_map< VertexPtr, unsigned int> vertexId;
+    for ( unsigned int i = 0, len = ( unsigned int ) vertices_.size( );
+          i < len; i++ )
     {
-      vertices_[i]->id( ) = i;
+      vertexId[ vertices_[i] ] = i;
       outStream << vertices_[i]->position( ).x( ) << " "
                 << vertices_[i]->position( ).y( ) << " "
                 << vertices_[i]->position( ).z( ) << "\n";
     }
     for( nlgeometry::FacetPtr facet: facets_ )
     {
-      outStream << "3 " << facet->vertex0( )->id( ) << " "
-                << facet->vertex1( )->id( ) << " "
-                << facet->vertex2( )->id( ) << "\n";
+      if ( facet->vertex3( ))
+        outStream << "4 " << vertexId[ facet->vertex0( )] << " "
+                  << vertexId[ facet->vertex1( )] << " "
+                  << vertexId[ facet->vertex2( )] << " "
+                  << vertexId[ facet->vertex3( )] << "\n";
+      else
+        outStream << "3 " << vertexId[ facet->vertex0( )] << " "
+                  << vertexId[ facet->vertex1( )] << " "
+                  << vertexId[ facet->vertex2( )] << "\n";
     }
 
+    vertexId.clear( );
     outStream.close();
   }