+++ Changelog.md
@@ -2,5 +2,10 @@
 
 ## git master
 
+	*[!67] Added support for vasculature mesh generation and visualization, added morphology extraction example.
+
+## 0.1.0
+
+	*[!66] nlrender and nlgenerator library refactored and commented, added tests and added examples.
 	*[!65] nlphysics library refactored and commented, added tests.
 	*[!63] nlgeometry library refactored and commented, added tests and added examples.